agent/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
python-dotenv>=1.0,<2.0
pipecat-ai[daily,openai,silero] @ git+https://github.com/pipecat-ai/pipecat.git@main 